fix for compile

+++ include/aquarius/basic_server.hpp
@@ -30,12 +30,12 @@ namespace aquarius
 		using callback_func = std::function<aquarius::awaitable<void>(std::shared_ptr<Session>)>;
 
 	public:
-		explicit basic_server(uint16_t port, int32_t io_service_pool_size, const std::string& name = {},
+		explicit basic_server(int32_t port, int32_t io_service_pool_size, const std::string& name = {},
 							  std::chrono::milliseconds global_time_dura = 30ms)
 			: io_service_pool_size_(io_service_pool_size)
 			, io_service_pool_(io_service_pool_size_)
 			, signals_(io_service_pool_.get_io_service(), SIGINT, SIGTERM)
-			, acceptor_(io_service_pool_.get_io_service(), detail::make_v4_endpoint(port))
+			, acceptor_(io_service_pool_.get_io_service(), detail::make_v4_endpoint(static_cast<uint16_t>(port)))
 			, server_name_(name)
 			, global_timer_(io_service_pool_.get_io_service(), global_time_dura)
 		{
